requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asttokens==3.0.0
attrs==25.1.0
bitsandbytes==0.45.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
datasets==3.3.2
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
executing==2.2.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.29.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.14.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
setuptools==75.8.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
trl==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.3
python-telegram-bot==1.3.0